+++ firebase-dataconnect/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- [changed] Internal refactor for managing Auth and App Check tokens
+  ([#7484](https://github.com/firebase/firebase-android-sdk/pull/7484),
+  [#7485](https://github.com/firebase/firebase-android-sdk/pull/7485))
+
 # 17.1.0
 
 - [fixed] Addressed minor reference documentation issues (#7399)